+++ cuiman/src/s2gos_client/api/client_mixin.py
@@ -2,7 +2,7 @@ from abc import ABC, abstractmethod
 from typing import Any
 
 from s2gos_common.models import ProcessDescription
-from s2gos_common.process import ExecutionRequest
+from s2gos_common.util.request import ExecutionRequest
 
 
 # noinspection PyShadowingBuiltins