+++ procodile/src/procodile/__init__.py
@@ -2,20 +2,24 @@
 #  Permissions are hereby granted under the terms of the Apache 2.0 License:
 #  https://opensource.org/license/apache-2-0.
 
-from .cli.cli import get_cli
+from importlib.metadata import version
+
+
+__version__ = version("procodile")
+
+from s2gos_common.util.request import ExecutionRequest
 from .job import Job, JobCancelledException, JobContext
 from .process import Process
 from .registry import ProcessRegistry
-from .request import ExecutionRequest
 
 """Processes development API."""
 
 __all__ = [
+    "__version__",
     "ExecutionRequest",
     "Job",
     "JobContext",
     "JobCancelledException",
     "ProcessRegistry",
     "Process",
-    "get_cli",
 ]